Upgrade: Bug fix for installation on Mac
Bug: When there is no /usr/local/madlib, installation fails due to the
non-existence of the folder when running preflight script

Fix: Detect whether the folder exists first in both preflight and
postflight scripts

+++ deploy/postflight.sh
@@ -15,5 +15,8 @@ ln -nsf /usr/local/madlib/Versions/$MADLIB_VERSION /usr/local/madlib/Current
 ln -nsf /usr/local/madlib/Current/bin /usr/local/madlib/bin
 ln -nsf /usr/local/madlib/Current/doc /usr/local/madlib/doc
 
-mv -f /usr/local/madlib/Versions.bak/* /usr/local/madlib/Versions/
-rm -rf /usr/local/madlib/Versions.bak
+if [ -d "/usr/local/madlib/Versions.bak" ]
+then
+    mv -f /usr/local/madlib/Versions.bak/* /usr/local/madlib/Versions/
+    rm -rf /usr/local/madlib/Versions.bak
+fi
